+++ src/common/layout/store-applayout.ts
@@ -3,23 +3,28 @@ import { create } from 'zustand';
 import { shallow } from 'zustand/shallow';
 
 interface AppLayoutStore {
-  // anchors - for externally closeable menus
-  drawerAnchor: HTMLElement | null;
-  menuAnchor: HTMLElement | null;
 
   // pluggable UI
   drawerItems: React.JSX.Element | null;
   centerItems: React.JSX.Element | null;
   menuItems: React.JSX.Element | null;
+
+  // anchors - for externally closeable menus
+  drawerAnchor: HTMLElement | null;
+  menuAnchor: HTMLElement | null;
+
 }
 
 const useAppLayoutStore = create<AppLayoutStore>()(
   () => ({
-    drawerAnchor: null,
-    menuAnchor: null,
+
     drawerItems: null,
     centerItems: null,
     menuItems: null,
+
+    drawerAnchor: null,
+    menuAnchor: null,
+
   }),
 );
 
@@ -36,16 +41,20 @@ export function useLayoutPluggable(centerItems: React.JSX.Element | null, drawer
 
 export function useLayoutComponents() {
   return useAppLayoutStore(state => ({
-    drawerAnchor: state.drawerAnchor,
-    menuAnchor: state.menuAnchor,
     drawerItems: state.drawerItems,
     centerItems: state.centerItems,
     menuItems: state.menuItems,
+    drawerAnchor: state.drawerAnchor,
+    menuAnchor: state.menuAnchor,
   }), shallow);
 }
 
 export function setLayoutDrawerAnchor(anchor: HTMLElement | null) {
   useAppLayoutStore.setState({ drawerAnchor: anchor });
+}
+
+export function closeLayoutDrawerMenu() {
+  useAppLayoutStore.setState({ drawerAnchor: null });
 }
 
 export function setLayoutMenuAnchor(anchor: HTMLElement | null) {
